be/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
httpx
python-dotenv
